Qt5 Workaround: Mask silent linking for windows
The line in the makefile that qmake generates:
LINKER := @echo linking $@ && g++
for CONFIG+=silent is broken on Windows.  This seems like probably
a bug in the version of mingw32-make shipped with Qt5.0.1 but
I'm unsure.  Until a solution is found this flag is masked for
windows.

This should be reverted and tested when the next release of Qt5
comes out.